Fix issue#33 Query with qdt.key.eq(literal) fails with operation error
If a query uses a literal as the parameter of an eq operation,
it is not returned when keys are obtained from the query handler.

The query handler is changed to check whether the parameter is a literal
and if so, returns the (escaped) literal. If not, it returns the (escaped)
value from the query parameters.