Fix for compiler warning for unused return value in ndb nodejs code
v8::Object::DefineOwnProperty() is marked V8_WARN_UNUSED_RESULT.
This results in a compiler warning that may escalated to an error.

In js_wrapper_macros.h fix the macro SET_RO_PROPERTY() to call
IsJust() on the v8::Maybe<bool> return value from DefineOwnProperty()
so that it is technically "used", though we still ignore the actual
bool result from the call.